Feat table to use effect stream
In this commit we change arguments of Table component to allow us using
streams.

Additionally: - Add `runtime.ts` with a runtime for browser - Add
`bindStreamToStore` util for Stream to Signal interop - Change test
cleanup to be done before a tests is run. So we can see result of the
last test in web interface - Add a test to see that our stream is
displayed correctly